+++ src/EnergyModelsCompliance.jl
@@ -7,5 +7,8 @@ const EMB = EnergyModelsBase
 const TS = TimeStruct
 
 include("compliance_access_fun.jl")
+include("compliance_ele.jl")
+
+export compliance_element
 
 end