+++ NEWS.md
@@ -2,4 +2,8 @@
 
 ## Version 0.1.0 (2025-03-xx)
 
-* Initial version of the package
+Initial version of the package:
+
+* Inclusion of compliance checks for new `Node`s and `Link`s
+  * Checks highlight whether a `Node` or `Link` can work with existing functions or have a new method.
+  * Checks mostly provide warnings, as errors are dependent on the chosen functions from `EnergyModelsBase`